Add files via upload
This was my main project for the robot that was used for most of the year during competitions.
I don't think a lot of the files are necessary such as makefile or the .mk files, but I uploaded just to be sure. 
This was made in the initial stages of vexcode so it doesn't use newer features like the controller setup that auto-creates a drive train in 'robot-config.cpp'.