Add files via upload
Used this file for testing the auton without any ability to use the driver controls.
Also, I tried to use a GUI for the pre-auton to set up the autonomous. However, I can't remember if it ever worked and I never used it for competitions (just preset the auton in the code).